backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
openai>=1.0.0
python-dotenv==0.19.1
requests==2.28.2 